FEATURE: renames file to api.service
