+++ libs/core/data-access/src/index.ts
@@ -1,5 +1,5 @@
 export * from './lib/auth.service';
 export * from './lib/batch.service';
-export * from './lib/batch';
+export * from './lib/batch-update.interface';
 export * from './lib/data.service';
 export * from './lib/local-storage.service';
