+++ libs/menu-matriarch/shared/data-access-api/src/lib/ingredient-update.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { BatchUpdate } from '@atocha/core/data-access';
 
-import { ApiService } from './batch.service';
+import { ApiService } from './api.service';
 import { Endpoint } from './endpoint.enum';
 
 @Injectable({
